Rename adapter files
