+++ src/core.ts
@@ -1,4 +1,4 @@
-export {createCMS} from './adapter/vanilla/VanillaCMS.js'
+export {createCMS} from './adapter/core/cms.js'
 export * from './core/Entry.js'
 export * from './core/Field.js'
 export * from './core/field/ListField.js'
