+++ src/core/Connection.ts
@@ -1,3 +1,4 @@
+import {HandleContext} from 'alinea/backend'
 import {Drafts} from 'alinea/backend/Drafts'
 import {History, Revision} from 'alinea/backend/History'
 import {PreviewInfo} from 'alinea/backend/Previews'
@@ -80,7 +81,7 @@ export namespace Connection {
     user?: User
     token?: string
   }
-  export interface Context extends AuthContext {
+  export interface Context extends AuthContext, HandleContext {
     logger: Logger
   }
   export const routes = {
